+++ Annex/Changes.md
@@ -1,5 +1,9 @@
 # Utopia Annex - List of Changes
 
+## v1.0.2 (in development)
+### Deprecations
+- Deprecated **RequestFailedException** in favor of another exception with the same name in **Utopia Disciple**
+
 ## v1.0.1 - 13.7.2021
 No changes in features, simply added support for the latest (breaking) changes in the 
 **Access** (v1.4) and **Flow** (v1.10) modules.