+++ Disciple/Changes.md
@@ -1,5 +1,9 @@
 # Utopia Disciple - List of Changes
 
+## v1.4.2 (in development)
+### New Features
+- Added a **RequestFailedException** class
+
 ## v1.4.1 - 13.7.2021
 No new features, only supports the latest (breaking) changes in the **Utopia Access** module.